Moves fillEdges methods into GenericDijkstraBidirection.
