+++ core/src/main/java/com/graphhopper/routing/AStarBidirection.java
@@ -144,61 +144,6 @@ public class AStarBidirection extends GenericDijkstraBidirection<AStarEntry> imp
     }
 
     @Override
-    boolean fillEdgesFrom() {
-        if (pqOpenSetFrom.isEmpty())
-            return false;
-
-        currFrom = pqOpenSetFrom.poll();
-        bestWeightMapOther = bestWeightMapTo;
-        fillEdges(currFrom, pqOpenSetFrom, bestWeightMapFrom, outEdgeExplorer, false);
-        visitedCountFrom++;
-        return true;
-    }
-
-    @Override
-    boolean fillEdgesTo() {
-        if (pqOpenSetTo.isEmpty())
-            return false;
-
-        currTo = pqOpenSetTo.poll();
-        bestWeightMapOther = bestWeightMapFrom;
-        fillEdges(currTo, pqOpenSetTo, bestWeightMapTo, inEdgeExplorer, true);
-        visitedCountTo++;
-        return true;
-    }
-
-    private void fillEdges(AStarEntry currEdge, PriorityQueue<AStarEntry> prioQueue,
-                           IntObjectMap<AStarEntry> bestWeightMap, EdgeExplorer explorer, boolean reverse) {
-        EdgeIterator iter = explorer.setBaseNode(currEdge.adjNode);
-        while (iter.next()) {
-            if (!accept(iter, currEdge.edge))
-                continue;
-
-            int traversalId = traversalMode.createTraversalId(iter, reverse);
-            if (!acceptTraversalId(traversalId, reverse)) {
-                continue;
-            }
-            double weight = calcWeight(iter, currEdge, reverse);
-            if (Double.isInfinite(weight))
-                continue;
-            AStarEntry entry = bestWeightMap.get(traversalId);
-            if (entry == null) {
-                entry = createEntry(iter, weight, currEdge, reverse);
-                bestWeightMap.put(traversalId, entry);
-                prioQueue.add(entry);
-            } else if (entry.getWeightOfVisitedPath() > weight) {
-                prioQueue.remove(entry);
-                updateEntry(entry, iter, weight, currEdge, reverse);
-                prioQueue.add(entry);
-            } else
-                continue;
-
-            if (updateBestPath)
-                updateBestPath(iter, entry, traversalId, reverse);
-        }
-    }
-
-    @Override
     protected AStarEntry createEntry(EdgeIteratorState iter, double weight, AStarEntry parent, boolean reverse) {
         AStarEntry entry;
         double currWeightToGoal = weightApprox.approximate(iter.getAdjNode(), reverse);
